added TimelineStore #76
